requirements.txt
langchain_openai 
langchain_core
langchain_community
python-dotenv
streamlit
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
wikipedia
langchain_groq
langchain_objectbox
ctransformers
sentence-transformers
pinecone-client
together